requirements.txt
albumentations==1.3.1
comet_ml==3.33.11
einops==0.7.0
FrEIA==0.2
imgaug==0.4.0
jsonargparse==4.25.0
kornia==0.6.9
matplotlib==3.8.0
numpy==1.26.0
omegaconf==2.3.0
opencv_python==4.8.1.78
opencv_python_headless==4.8.1.78
openvino==2023.1.0
pandas==2.1.1
Pillow==10.0.1
pytorch_lightning==1.9.1
Requests==2.31.0
scikit_learn==1.3.1
scipy==1.11.3
setuptools==65.5.1
tifffile==2023.9.26
torchvision==0.16.0
tqdm==4.66.1
urllib3==2.0.6
pytorch_lightning==1.9.1
timm==0.9.7
tensorboard==2.14.1